+++ src/Nethermind/Nethermind.Core/JsonConverters/AddressConverter.cs
@@ -1,16 +1,13 @@
-// SPDX-FileCopyrightText: 2022 Demerzel Solutions Limited
+// SPDX-FileCopyrightText: 2024 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using System;
 using System.Runtime.CompilerServices;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-
-using Nethermind.Core;
 using Nethermind.Core.Extensions;
-using Nethermind.Core.JsonConverters;
 
-namespace Nethermind.Serialization.Json;
+namespace Nethermind.Core.JsonConverters;
 
 public class AddressConverter : JsonConverter<Address>
 {
@@ -19,7 +16,7 @@ public class AddressConverter : JsonConverter<Address>
         Type typeToConvert,
         JsonSerializerOptions options)
     {
-        byte[]? bytes = ByteArrayConverter.Convert(ref reader);
+        var bytes = ByteArrayConverter.Convert(ref reader);
         return bytes is null ? null : new Address(bytes);
     }
